node.label {tidytree}	R Documentation

extract the node label of phylo, treedata or tbl_tree

Description

extract the node label of phylo, treedata or tbl_tree

Usage

node.label(x, node = "internal", ...)

Arguments

x
object, should be one of treedata,phylo or tbl_tree.

node
character, to extract which type node label, default is internal, should be one of internal, external, all, tip.

...
additional parameters.

Value

label character vector.

[Package tidytree version 0.4.6 Index]