td-label-assign {tidytree}	R Documentation

the tip or internal node label assign of tbl_tree phylo and treedata

Description

the tip or internal node label assign of tbl_tree phylo and treedata

Usage

tip.label(x) <- value

node.label(x) <- value

## S3 replacement method for class 'phylo'
node.label(x) <- value

## S3 replacement method for class 'treedata'
node.label(x) <- value

## S3 replacement method for class 'tbl_tree'
node.label(x) <- value

## S3 replacement method for class 'phylo'
tip.label(x) <- value

## S3 replacement method for class 'treedata'
tip.label(x) <- value

## S3 replacement method for class 'tbl_tree'
tip.label(x) <- value

Arguments

x
object, should be one of tbl_tree, phylo or treedata

value
character, the character vector

[Package tidytree version 0.4.6 Index]