tip.label {tidytree}	R Documentation

extract the tip label of phylo treedata or tbl_tree

Description

extract the tip label of phylo treedata or tbl_tree

Usage

tip.label(x, ...)

Arguments

x
object, should be one of treedata,phylo or tbl_tree.

...
additional parameters.

[Package tidytree version 0.4.6 Index]